% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showMixed.R
\name{fixedCoef}
\alias{fixedCoef}
\title{Extract Model Fixed Coefficients}
\usage{
fixedCoef(object, names = TRUE)
}
\arguments{
\item{object}{is an object of class \sQuote{mixed}, created by \code{readCoef}.}

\item{names}{is logical for whether to print solution lables (TRUE) or not (FALSE). Default is TRUE.}
}
\value{
Numeric vector with solutions labled based on model formula and data.
}
\description{
Interacts with an object of class \sQuote{mixed} to extract all fixed factors along with any covariates fit in the model.
}
