% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showMixed.R
\name{getLabel}
\alias{getLabel}
\title{Get Labels of all Solutions as Stored in Mixed Model Object}
\usage{
getLabel(mixedOBJ, optn = "all")
}
\arguments{
\item{mixedOBJ}{an object of class \sQuote{mixed}, created by \code{readCoef}.}

\item{optn}{default option is to get labels for \sQuote{all} variable solutions in the mixed model; other options are \sQuote{fixed} and \sQuote{random}.}
}
\description{
Get Labels of all Solutions as Stored in Mixed Model Object
}
