% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showMixed.R
\name{getRel}
\alias{getRel}
\title{Extract Reliabilities of Random Effects}
\usage{
getRel(object, names = TRUE, list = TRUE)
}
\arguments{
\item{object}{is an object of class \sQuote{mixed}, created by \code{readCoef}.}

\item{names}{is logical for whether to print solution lables (TRUE) or not (FALSE). Default is TRUE.}

\item{list}{logical for whether to extract reliabilities in the form of a list (default) or as a numeric vector by setting \code{list = FALSE}.}
}
\value{
Numeric vector with reliabilities labled based on model formula and data.
}
\description{
Extract reliabilities of random efects from an object of class \sQuote{mixed}.
}
\details{
Used in situations where it is important to package together reliabilities of all random effects in one numeric vector with matching lables; this is done by setting \code{list = FALSE}. In cases where reliabilities of a particular random effect are needed, extract reliabilities as a list using the option \code{list = TRUE}.
}
