% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showMixed.R
\name{mixedPredict}
\alias{mixedPredict}
\title{Predict Function using an Object of Class \sQuote{mixed}.}
\usage{
mixedPredict(mixedOBJ, newdata = eval(parse(text =
  as.character(mixedOBJ@dataName))))
}
\arguments{
\item{mixedOBJ}{is an object of class \sQuote{mixed}, created by \code{readCoef}.}

\item{newdata}{data frame with new data to predict, default is to use training data frame associated with the mixed object}
}
\value{
Numeric vector with predictions of length equal to number of rows in newdata
}
\description{
Predict Function using an Object of Class \sQuote{mixed}.
}
\details{
In the case of \code{TM = TRUE}, predictions are calculated on the uunderlying continuous scale with no transformation ot the probability scale yet.
}
\note{
New to gep 0.1.1 (added January 2019)
}
