% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showMixed.R
\name{mixedSummary}
\alias{mixedSummary}
\title{Summary info for a mixed object}
\usage{
mixedSummary(object, random = FALSE)
}
\arguments{
\item{object}{is an object of class \sQuote{mixed}, created by \code{readCoef}.}

\item{random}{is logical for whether to print random solutions. Default is FALSE.}
}
\value{
None. Information is printed to standard output.
}
\description{
Interacts with an object of class \sQuote{mixed} to extract model info and random coefficients if specified.
}
