% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showMixed.R
\name{randomCoef}
\alias{randomCoef}
\title{Extract Model Random Coefficients}
\usage{
randomCoef(object, names = TRUE, list = TRUE)
}
\arguments{
\item{object}{is an object of class \sQuote{mixed}, created by \code{readCoef}.}

\item{names}{is logical for whether to print solution lables (TRUE) or not (FALSE). Default is TRUE.}

\item{list}{logical for whether to extract the solutions in the form of a list (default) or as a numeric vector by setting \code{list = FALSE}.}
}
\value{
Numeric vector with solutions labled based on model formula and data.
}
\description{
Extract solutions of random factors from an object of class \sQuote{mixed}.
}
\details{
Used in situations where it is important to package together solutions of all random effects in one numeric vector with matching lables; this is done by setting \code{list = FALSE}. In cases where a particular random effect is needed, extract random solutions as a list using the option \code{list = TRUE}.
}
