% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCoef.R
\name{readCoef}
\alias{readCoef}
\title{Read Mixed Model Solutions from disk and Package in an Object of Class mixed}
\usage{
readCoef(sol.file = "solFile")
}
\arguments{
\item{sol.file}{path and name of sotored mixed model solutions on disk. Default sol.file = "./solFile".}
}
\description{
Function \code{mixed} saves coefficients on disk. If more work on coefficients is needed, read them back from disk and package them in an objcet of class \sQuote{mixed}.
}
\details{
WARNING: Do not alter the original copy of solutions file or \code{readCoef} will fail.
}
