% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.sparse.R
\name{readSparse}
\alias{readSparse}
\title{Function to read sparse-stored matrix on disc into an R list}
\usage{
readSparse(file = "")
}
\arguments{
\item{file}{is a character string for the file name and directory path where the sparse-stored matrix is stored on disc.}
}
\value{
A list including 3 vectors, 'ia' for rows, 'ja' for columns, and 'a'
  for elements, where \code{length(sparseMatrix$ia)} = the order of the matrix and \code{length(sparseMatrix$ja)}
  = the number of nonzeros; and 2 scalers for the matrix order and number of non-zero elements.
}
\description{
Function to read sparse-stored matrix on disc into an R list
}
\details{
Read sparse-stored matrix on disc in the supplied file name
}
