% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store.full.R
\name{storeFull}
\alias{storeFull}
\title{Transform sparse-stored to full-stored matrix}
\usage{
storeFull(sparseMatrix)
}
\arguments{
\item{sparseMatrix}{is a list created by function \code{readSparse}.}
}
\value{
Full-stored matrix with dimensions = order of input sparse matrix.
}
\description{
Transform sparse-stored to full-stored matrix
}
\details{
The input sparse matrix is a list of 3 vectors, 'ia' for rows, 'ja' for columns, and 'a'
  for elements, where \code{length(sparseMatrix$ia)} = the order of the matrix and \code{length(sparseMatrix$ja)}
  = the number of nonzeros.
}
